changed atomindices to int instead of size_t

+++ include/intraNonBonded/intraNonBondedContainer.hpp
@@ -18,17 +18,17 @@ namespace intraNonBonded
 class intraNonBonded::IntraNonBondedContainer
 {
   private:
-    size_t                           _molType;
-    std::vector<std::vector<size_t>> _atomIndices;
+    size_t                        _molType;
+    std::vector<std::vector<int>> _atomIndices;
 
   public:
-    IntraNonBondedContainer(const size_t molType, const std::vector<std::vector<size_t>> &atomIndices)
+    IntraNonBondedContainer(const size_t molType, const std::vector<std::vector<int>> &atomIndices)
         : _molType(molType), _atomIndices(atomIndices)
     {
     }
 
-    [[nodiscard]] size_t                           getMolType() const { return _molType; }
-    [[nodiscard]] std::vector<std::vector<size_t>> getAtomIndices() const { return _atomIndices; }
+    [[nodiscard]] size_t                        getMolType() const { return _molType; }
+    [[nodiscard]] std::vector<std::vector<int>> getAtomIndices() const { return _atomIndices; }
 };
 
 #endif   // _INTRA_NON_BONDED_CONTAINER_HPP_
